+++ pcs/lib/cib/resource/primitive.py
@@ -70,6 +70,7 @@ def create(
         report_processor,
         raw_operation_list,
         resource_agent.get_default_actions() if use_default_operations else [],
+        [operation["name"] for operation in resource_agent.get_actions()],
         allow_invalid=allow_invalid_operation,
     )